Use REST URI as OFC resource identifier
In Quantum resource relationship is not limited inside a tenant.
E.g., a non-owner tenant can create a port on a public network.
To deal with it the provider layer should not be aware of tenants
each resource belongs to even when it has a kind of tenant concept.
This commit changes ofc_manager to pass a parent resource for resource
creation and identify a resouce by REST URI used to access OFC resources.
It decouples Quantum resource access model from OFC resource models.