plugin/nec: Make sure resources on OFC is globally unique.
Some OpenFlow controller (OFC) implmenetation generates a network ID
unique inside a tenant. On the other hand, ID mappng table in NEC plugin
expects ID on OFC is globally unique. As a result creation of a network
will fail. To fix it, quantum network ID is passed as ID on OFC.
The same issue happens in port creation too.

Change-Id: Iad1961ee2469b547598e87e63a4a3dc4ac9c6e97