plugin/nec: Ensure ID on PFC is up to 31 chars
Some version of NEC ProgrammableFlow OpenFlow controller, one of
the OpenFlow controllers (OFCs), accepts the string up to 31 chars
as a network ID. NEC plugin now uses Quantum UUID as ID on OFC,
so this leads to failure on network creation. This commit changes
the plugin to generate a 31-char ID on OFC to keep its uniqueness.
The same thing can be said to port ID.

Similarly, the description field must be a string up to 127 chars
and consist of [A-Za-z0-9_] characters. This change also fixes it.

Change-Id: I0a879f5158a6ac0ce332e3698d42dd38d4283a60