vrouter: call ofc_delete_tenant when OFC tenant has no resource.
Previously ofc_delete_tenant is called when all resources in
quantum tenant are deleted. This commit changes the behavior
to call ofc_delete_tenant when all resources on OFC are deleted.
In addition, it ensures ofc_delete_tenant is not called when
a corresponding tenant does not exist on OFC.

Change-Id: If92c878ead34890ce8a433f521d7677e0080177d